Log client query duration and search_path
adds a new configuration option, log_min_duration, which if set to 0,
enables logging of all query durations. If set to a number > 0, then
it logs any query who's duration is longer than log_min_duration
milliseconds. If the value of log_min_duration is < 0, no logging
of query durations happens.

Note: in addition to duration, the stored search_path is also included
in the log output (if set).